:recycle: Extract issueList composable
